requirements.txt
aiogram==3.1.1
asyncpg==0.27.0
Flask==2.3.2
web3==6.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
